backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
langchain==0.3.10
langchain-anthropic==0.3.2
langchain-community==0.3.10
langchain-openai==0.2.8
langgraph==0.2.55
pydantic==2.10.4
python-dotenv==1.0.1
psycopg2-binary==2.9.10
sqlalchemy==2.0.36
alembic==1.14.0
aiokafka==0.11.0
websockets==13.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
ipython==8.25.0
confluent-kafka==2.5.0
faker==32.0.0
chromadb==0.5.23
httpx==0.27.0
